# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Trade Your Exp.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "Авторизирует пользователя по введенным данным.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно вошел",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Создание пользователя с заполнением требуемых полей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создание пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "409": {
                        "description": "Пользователь уже существует",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/skills": {
            "get": {
                "description": "Вывод полного списка навыков.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Вывести все скиллы",
                "responses": {
                    "200": {
                        "description": "Вывод пользовательских скиллов",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание нового запроса с описанием",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Добавление запроса с дополнительным описанием",
                "parameters": [
                    {
                        "description": "Данные запроса навыка",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SkillFull"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Запрос навыка успешно создан",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "409": {
                        "description": "Конфликт при создании запроса",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/skills/desc": {
            "get": {
                "description": "Вывод полного дополнительных описаний",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Вывести все доп.описания",
                "responses": {
                    "200": {
                        "description": "Вывод списка доп. описаний",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/skills/desc/{id}": {
            "get": {
                "description": "Выводит описание по персональному skill_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Вывод описания по skill_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Вывод описания по skill_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вывод пользователя",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат Skill ID",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/skills/filter/{search}": {
            "get": {
                "description": "Вывод полного списка скиллов по переданной строке.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Вывести все скиллы по вхождению в ключевую строку.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ключевая строка поиска",
                        "name": "search",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вывод пользовательских скиллов",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "404": {
                        "description": "Категория не найдена",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/skills/{category}": {
            "get": {
                "description": "Вывод полного списка скиллов по категории.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Вывести все скиллы по категории.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Категория скилла",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вывод пользовательских скиллов",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "404": {
                        "description": "Категория не найдена",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/skills/{id}": {
            "delete": {
                "description": "Удаляет запрос навыка по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Удалить запрос",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID запроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запрос навыка успешно удален",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "404": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    },
                    "500": {
                        "description": "Запрос не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseApi"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "description": "Обновляет данные в базе данных по его UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Обновить данные",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно обновлен"
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет пользователя из базы данных по его UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Удалить пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно удален"
                    },
                    "404": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "role",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "manager",
                        "admin",
                        "viewer"
                    ]
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ResponseApi": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "[!] доработать код ошибки [!]",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Information message"
                },
                "request_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "result": {},
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SkillFull": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "media": {
                    "type": "string"
                },
                "skill": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "description": "omitempty чтобы не светить в API",
                    "type": "string"
                },
                "socialLink": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}